Use checkboxes instead of a dropdown
This requires more more. Right now, only the options can be saved but meta boxes and the front-end won't work with the stored array. Moreover, backwards-compatibility must be ensured. To ensure backwards-compatibility, we can either map an already stored option value to the according checkboxes, or we introduce a new option name that takes precedence over the old option